Add severity to stack trace
