+++ stack.go
@@ -1,5 +1,22 @@
 package sterrors
 
+import (
+	"runtime"
+	"strings"
+)
+
+type Caller struct {
+	FuncName string `json:"funcName,omitempty"`
+	File     string `json:"file,omitempty"`
+	Line     int    `json:"line,omitempty"`
+}
+
+type CallStackEntry struct {
+	ErrMessage string   `json:"msg,omitempty"`
+	Severity   Severity `json:"severity,omitempty"`
+	Caller     Caller   `json:"caller,omitempty"`
+}
+
 // CallStack returns the callstack
 // It adds CallStack entries recursively based on error causes
 // as long they also implement the Error interface
@@ -14,7 +31,7 @@ func CallStack(err error) []CallStackEntry {
 		return append(stack, CallStackEntry{ErrMessage: err.Error()})
 	}
 
-	res := []CallStackEntry{{ErrMessage: e.Error(), Caller: e.Caller()}}
+	res := []CallStackEntry{{ErrMessage: e.Error(), Caller: e.Caller(), Severity: e.Severity()}}
 
 	if e.Cause() == nil {
 		return res
@@ -23,4 +40,15 @@ func CallStack(err error) []CallStackEntry {
 	res = append(res, CallStack(e.Cause())...)
 
 	return res
+}
+
+func caller() Caller {
+	pc, file, line, _ := runtime.Caller(2)
+	details := runtime.FuncForPC(pc)
+	nameSegments := strings.Split(details.Name(), "/")
+	funcName := details.Name()
+	if len(nameSegments) > 0 {
+		funcName = nameSegments[len(nameSegments)-1]
+	}
+	return Caller{File: file, Line: line, FuncName: funcName}
 }
